+++ example/lib/features/dependency_wheel/dependency_wheel_screen.dart
@@ -1,7 +1,5 @@
 import 'package:example/features/dependency_wheel/screens/commercial_balance_dependency_wheel_screen.dart';
 import 'package:example/features/dependency_wheel/screens/import_export_dependency_wheel_screen.dart';
-import 'package:example/features/treemap_chart/screens/basic_treemap_chart_screen.dart';
-import 'package:example/features/treemap_chart/screens/category_treemap_chart_screen.dart';
 import 'package:flutter/material.dart';
 
 class DependencyWheelScreen extends StatelessWidget {
@@ -14,29 +12,29 @@ class DependencyWheelScreen extends StatelessWidget {
       body: ListView(
         children: [
           ListTile(
-            title: Text("Sells of iPhones and Androids"),
-            subtitle: Text(""),
+            title: const Text("Sells of iPhones and Androids"),
+            subtitle: const Text(""),
             onTap: () {
               Navigator.of(context).push(
                 MaterialPageRoute(
-                  builder: (context) => AndroidIphoneDependencyWheelScreen(),
+                  builder: (context) => const AndroidIphoneDependencyWheelScreen(),
                 ),
               );
             },
           ),
-          Divider(),
+          const Divider(),
           ListTile(
-            title: Text("Commercial Balande between different countries"),
-            subtitle: Text(""),
+            title: const Text("Commercial Balande between different countries"),
+            subtitle: const Text(""),
             onTap: () {
               Navigator.of(context).push(
                 MaterialPageRoute(
-                  builder: (context) => ImportExportDependencyWheelScreen(),
+                  builder: (context) => const ImportExportDependencyWheelScreen(),
                 ),
               );
             },
           ),
-          Divider(),
+          const Divider(),
         ],
       ),
     );