+++ example/lib/features/dot_plot/screens/basic_dot_plot_screen.dart
@@ -1,4 +1,3 @@
-import 'package:example/features/bubble_chart/bubble_chart_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:pretty_charts/pretty_charts.dart';