+++ example/lib/features/scatter_plot/screens/basic_scatter_plot.dart
@@ -1,4 +1,3 @@
-import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:pretty_charts/pretty_charts.dart';