+++ example/lib/features/ternary_plot/screens/basic_ternary_plot_screen.dart
@@ -19,7 +19,7 @@ class BasicTernaryPlotScreen extends StatelessWidget {
               leftAxesTitle: "",
               rightAxesTitle: "",
             ),
-            data: [],
+            data: const [],
           ),
         ),
       ),