+++ lib/src/plots/scatter_plot/scatter_plot.dart
@@ -72,20 +72,23 @@ class _ScatterPlotState extends State<ScatterPlot>
           _offset = offset;
         });
       },
-      child: ClipRect(
-        child: CustomPaint(
-          painter: ScatterPlotPainter(
-            scaleFactor: _scaleFactor,
-            animationProgress: _progressAnimation.value,
-            offset: _offset,
-            data: widget.data,
-            colorMap: widget.colorMap ?? pastel1,
-            axes: widget.axes,
-          ),
-          foregroundPainter: PlotFrameworkPainter(
-            axes: widget.axes,
-            scaleFactor: _scaleFactor,
-            offset: _offset,
+      child: LayoutBuilder(
+        builder: (context, constraints) => ClipRect(
+          child: CustomPaint(
+            size: Size(constraints.maxWidth, constraints.maxHeight),
+            painter: ScatterPlotPainter(
+              scaleFactor: _scaleFactor,
+              animationProgress: _progressAnimation.value,
+              offset: _offset,
+              data: widget.data,
+              colorMap: widget.colorMap ?? pastel1,
+              axes: widget.axes,
+            ),
+            foregroundPainter: PlotFrameworkPainter(
+              axes: widget.axes,
+              scaleFactor: _scaleFactor,
+              offset: _offset,
+            ),
           ),
         ),
       ),