+++ lib/src/plots/treemap_chart/treemap_chart_data.dart
@@ -1,4 +1,3 @@
-import 'package:pretty_charts/pretty_charts.dart';
 
 class TreeMapChartData {
   final String name;