+++ CHANGELOG.md
@@ -1,3 +1,19 @@
-## 0.0.1
+## 0.0.1-dev.1
 
-* TODO: Describe initial release.
+First dev release of the Pretty Charts package
+
+It can render :
+
+- Pie Plot
+- Line Plot
+- Bar Plot
+- Bubble Chart
+- Scatter Plot
+- Dependency Wheel
+- Ternary Plot
+- Treemap Chart
+- Dot Chart
+- Contour Plot
+
+The first rendering is animated excepting for the contour plot.
+There's no interaction yet.